refactor : common으로  파일 이전 DownloadImageDto, ProfileImageDto
