+++ src/main/java/com/monari/monariback/common/dto/ProfileImageDto.java
@@ -1,0 +1,9 @@
+package com.monari.monariback.common.dto;
+
+public record ProfileImageDto(
+		String profileImageKey
+) {
+	public static ProfileImageDto from(String profileImageKey) {
+		return new ProfileImageDto(profileImageKey);
+	}
+}
