+++ src/main/java/com/monari/monariback/teacher/controller/TeacherController.java
@@ -8,12 +8,15 @@ import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
 
 import com.monari.monariback.auth.aop.Auth;
 import com.monari.monariback.auth.aop.OnlyTeacher;
 import com.monari.monariback.auth.entity.Accessor;
 import com.monari.monariback.teacher.dto.request.TeacherUpdateRequest;
+import com.monari.monariback.teacher.dto.response.TeacherProfileImageResponse;
 import com.monari.monariback.teacher.dto.response.TeacherResponse;
 import com.monari.monariback.teacher.service.TeacherService;
 
@@ -55,6 +58,19 @@ public class TeacherController {
 		return ResponseEntity.ok().body(
 				TeacherResponse.from(
 						teacherService.updateProfile(accessor, request)
+				)
+		);
+	}
+
+	@OnlyTeacher
+	@PatchMapping("/me")
+	public ResponseEntity<TeacherProfileImageResponse> updateProfile(
+			@Auth Accessor accessor,
+			@RequestPart("file") MultipartFile file
+	) {
+		return ResponseEntity.ok().body(
+				TeacherProfileImageResponse.from(
+						teacherService.updateProfileImage(accessor, file)
 				)
 		);
 	}
