+++ src/main/java/com/monari/monariback/teacher/dto/response/TeacherProfileImageResponse.java
@@ -1,0 +1,9 @@
+package com.monari.monariback.teacher.dto.response;
+
+public record TeacherProfileImageResponse(
+		String profileImageKey
+) {
+	public static TeacherProfileImageResponse from(String profileImageKey) {
+		return new TeacherProfileImageResponse(profileImageKey);
+	}
+}
